PMIC
Switching to a PMIC topology since its a better solution for powering
the diverse set of hardware on the ZyncMV. The PMIC can be controlled by
the processor and shut down the device like a regular SBC. The PMIC als
has all the appropraite voltages for the memory and enough current to
support it. Additional regulartors still mayb required for supplying the
image sensor and most defintiely the light sources.