USB
This commit contains mostly work done to the USB sections. Theres a FTDI
chip for USB coms and programmign and an USB transceiver/PHY for usb
perfipherals or daisychaining devices.  Also the termiantion resistors
have been removed from the memory since termination can be set on DDR
pins of the FPGA with verilog.